[usdMaya] check the bbox env setting before USD stage valid-ness in UsdMayaProxyShape::isBounded()
Checking the PIXMAYA_ENABLE_BOUNDING_BOX_MODE TfEnvSetting first should
allow the conditional to short-circuit sooner, since it is almost always
disabled, and doing so avoids incurring the potential Maya DG evaluation in
isStageValid().

There doesn't seem to be any evidence so far that this makes a significant
difference in performance, but it can't hurt either.

(Internal change: 1883863)