Send notices for all prims/properties when removing inert subtree.
When removing an inert spec, Sdf would previously only send
a corresponding 'inert spec removal' notice for the spec
that was explicitly removed. Downstream clients like Pcp
expected this notice to mean that only the removed spec was
affected. However, this was incorrect: an inert spec may
have inert prim or property children. So, removing an
inert spec may result in the removal of an entire subtree
of inert specs.

Fixing up the downstream clients would be difficult, as
a lot of infrastructure has been built up on the assumption
that notices about inert specs do not require recursing
down subtrees of prims and properties. So instead, Sdf
now explicitly removes each spec in an inert subtree
from the bottom-up. This is similar to existing behavior
in SdfLayer::_SetData and ensures that the layer change
notice includes each inert spec being removed.

(Internal change: 1884135)