[pxrUsd/pxrUsdMayaGL] clean up draw registry plugin registration
The classification scheme for draw overrides for shapes is
"drawdb/geometry/<library name>/<shape type name>", so this change conforms
the classication for UsdMayaProxyDrawOverride to that. Also, the pxrUsd
plugin is the thing performing the registration in Maya's draw registry, so
it should own its registrant ID.

Finally, this change also prevents copying and assignment for
UsdMayaProxyDrawOverride.

(Internal change: 1890434)