[usdImagingGL] Fix for UsdImaging's TextureId not accounting for origin.
The texture id is supposed to encapsulate a unique texture object and sampler object.  Code was recently added such that a texture now has an origin (either lower-left or lower-right), which is resolved when the texture is loaded and results in a unique texture object.

Which origin to use is based on the shader node's id that is used to reference the texture.

Changes:
- Refactored the code to calculate the texture origin for a shader node into a static subroutine.
- When computing the texture id, also hash in the origin.

(Internal change: 1890486)