[HdSt] Fix garbage collection on buffer migration; MarkShaderBindingsDirty => MarkBatchesDirty; fix MarkAllRprimsDirty.
We were avoiding garbage collection in a bunch of places in hydra because it was previously really expensive: it called MarkCollectionsDirty, which would rebuild dirty lists and batches.

I've re-instated garbage collection on buffer migration in HdStMesh and removed the "MarkAllCollectionsDirty" from the garbage collection phase.  This involved chasing down some other invalidation bugs that were masked: for example, MarkAllRprimsDirty() wasn't setting the varying bit, so the ptex transition code wasn't firing correctly.  Some of the SetGarbageCollectionNeeded call-sites now need a MarkBatchesDirty call as well.

MarkBatchesDirty is a more appropriate name for MarkShaderBindingsDirty, since the shader bindings version is set and read as a trigger to re-validate aggregation, which includes both shaders and buffers.

(Internal change: 1890963)