Fix extra field padding issues in .usdz files
UsdZipFileWriter writes padding bytes to the "extra"
section of each file's local header to ensure data is
aligned to 64-byte boundaries. However, it was not
writing the 4-byte header in this section as required
by the zip file specification (section 4.5). In addition,
the "extra" section wasn't being replicated to the file's
central directory record(*). Both issues are now fixed.

This issue does not affect the ability to read .usdz
files using USD itself. However, it does mean that
previously-created .usdz files do not fully conform
to the zip file specification, which could lead to
problems  reading these files with external zip utilities
and libraries. In practice, it appears that many of these
utilities are unaffected as they skip over the extra field
data entirely, but the missing header could still be
problematic for tools that require strict conformance.

This change also fixes an inefficiency where padding
bytes were written to the .usdz file even if the file data
was already aligned properly.

(*) The spec isn't explicit about requiring the "extra"
section to be replicated in the central record, but it
appears to be the case based on a cursory look at
some other zip utilities.

(Internal change: 1891606)