Allow SdfCopySpec to copy prims to/from variants
SdfCopySpec now properly handles copying prim specs
to variants (e.g. /Foo to /Bar{x=a}) and vice-versa.
This would previously result in an error if the
destination spec did not already exist.

One tweak to the copying behavior: when copying a
variant to a prim, the specifier and typename from
the variant's owning prim will be used.

(Internal change: 1892515)