[Gf] Add common raw pointer access for GfMatrix* classes.
There are a few places in the tool chain where it is necessary to serialize GfVec* and GfMatrix* to and from an array of values.  While both sets of classes provided methods to access the underlying raw pointer, there are some subtle differences between the interfaces and there wasn't a single common API that could be used for both.    This made it more challenging to have a templated serialization function that could be used for both of these groups of classes.

After discussion data() was chosen to be the common API.  GfVec already implements a const and non-const version of data().  GfMatrix* did not implement this method, this change doesn't affect any existing users.

Changes:
- Added Code Gen templates for const and non-const data() methods.
- Updated the resulted code gen files.

(Internal change: 1892740)