Remove copy-list-initialization in TfEnvSetting
A crash in TfEnvSetting was reported in pull request #549
that appeared to be caused by the pointer in the
TfEnvSetting's std::atomic value holder being reset to 0
*after* it had been initialized in the corresponding
registry function.

Experiments on godbolt and debug builds yielded some
interesting results. On gcc, clang and MSVC 2015, the
previous code with the explicit copy-list-initialization
caused the std::atomic to be zero-initialized before the
registry function was run, so the crash did not occur.
On MSVC 2017, however, it appears that the std::atomic
c'tor that takes an initial value is invoked instead,
potentially after the registry function is run.

The fix for this as suggested in the pull request
is to simply remove the copy-list-initialization and
rely on standard zero initialization for static
variables. Per godbolt, this seems to give us the
desired behavior in all cases, including MSVC 2017.
On gcc with optimizations disabled (-O0), this change
causes static initializers to be generated for the
std::atomic variables, which theoretically could
lead to the same problem as above. However, from
looking at the assembly and some ad-hoc testing in
gdb, it appears that these initializers wind up
having no effect.

Thanks to Mark Tucker for reporting the issue and
providing helpful debugging info!

Fixes #549

(Internal change: 1893406)