Force usdedit to read-only mode for usdz files
Running usdedit on a layer in a .usdz file and making
modifications in the editor would previously cause
errors. This is because layers in .usdz files (or any
other package file format) cannot be modified directly
via the Sdf API.

To avoid this confusion, usdedit will now emit a warning
when it's used to open a layer in a .usdz file and force
read-only mode. This change also adds support for
package-relative paths to usdedit, so that a user can
run usdedit on an arbitrary file in a package file.

(Internal change: 1893494)