[Hd, UsdImaging] Removed Repr from GetRenderTag api.
Some of our tools have the options to display mesh vertices as points as well as the mesh.  This used to be done via a separate collection with the points repr, using the render tag to control the visibility.

As the render tag is used as an early filter on the prim lists (in dirty list), fetching the render tag is performance sensitive and due to the interaction with repr, added an extra dimension to the problem of caching the render tag.

With the change over to HdReprSelector, their was no longer any need for the additional points collection, as well as using the render tag system to have a dependency on Repr.

Therefore, this change removes repr from the GetRenderTag api.  Leaving additional optimization for future work.

Changes:
- Removed the reprSelector argument from HdSceneDelegate::GetRenderTag().
- Updated all delegate implementations with this API change (done used this variable anymore).
- Removed reprSelector from the calling functions that got the render tag from the scene delegate.

(Internal change: 1893690)