[Hdx, Maya] Added a new hit resolution method for use when ray picking
For resolving the ID render for ray picks, HdxIntersector::Result provides ResolveNearest, which walks through all the draw target pixels and returns the hit nearest to the camera. This leads to picking woes in, for example, face picking, wherein, instead of picking the face the mouse is on, we return the another that's closer to the camera (that wasn't culled during the ID render).
With mouse rollovers, or fine geometry, or zoomed-in cameras, this becomes very evident.

This CL attempts to address this with the following changes:
- Added a function ResolveNearestToCenter to return the (first) valid hit closest to the pick point. This essentially means we sample the ID render draw target from the center outwards in concentric rectangles/squares.
- Renamed the earlier ResolveNearest method to ResolveNearestToCamera.
- Updated cleints (usdview and maya) to use it.

(Internal change: 1893861)