[usdMaya/pxrUsdMayaGL] introduce pxrHdImagingShape node for better Hydra batched drawing
This change adds a pxrHdImagingShape with accompanying MPxSurfaceShapeUI
and MPxDrawOverride subclasses for invoking Hydra draws in response to Maya
callbacks.

This shape node is intended to take responsibility for the bulk of the
actual drawing of Hydra-imaged shapes and cut down on the overhead of
having to manage many callbacks from potentially many shapes. Before this
node, every shape would translate its Maya viewport draw callback into a
call to Draw() on the batch renderer. The batch renderer then had to do
some bookkeeping to determine when to actually invoke Hydra (usually the
first shape to call Draw()), and when to simply passthrough (usually every
Draw() call after the first). Instead, the batch renderer now always invokes
Hydra when the render params specify that the shape should be drawn since
typically all shapes will have drawShape set to false except for the one
pxrHdImagingShape in the scene that will have it set to true.

Note that all Hydra-imaged shapes will likely still be calling the batch
renderer's Draw() method for now. This is because the drawing of bounding
boxes is not yet handled by Hydra and is instead done internally by the
batch renderer on a per-shape basis. Ideally that would eventually get
folded into Hydra as well, and in that case we may not even need to
register draw callbacks for MPxDrawOverride subclasses, further reducing
the Maya overhead.

(Internal change: 1893936)