Put USD include directories first in search order
Ensure include directories are set up so that the directories
containing the USD headers for the current build always come
before the include directories for external dependencies.

When a user had installed USD into the same location as
the external dependencies (which is what build_usd.py does
by default), the old ordering would cause the build to prefer
the previously-installed headers if USD was rebuilt into
that location. This could cause the build to fail if those
headers conflicted with those that were in the source tree.

Fixes #83

Fixes #83

(Internal change: 1893985)