Added an Sdr parser plugin in UsdShade (UsdShadeShaderDefParserPlugin) for parsing shader definitions represented by shader prims authored using the UsdShade schema/API in a USD file.
Renamed shaderMetadata to sdrMetadata, so it can apply to shader properties
(inputs and outputs) in addition to shader prims. Added API for authoring and
querying sdrMetadata on UsdShadeInput and UsdShadeOutput. sdrMetadata
authored on shader properties in shader definitions are recorded on the
corresponding SdrShaderProperty in the SdrRegistry.

Added API UsdShadeShaderDefUtils::GetNodeDiscoveryResults(), which returns the
list of NdrNodeDiscoveryResult objects that must be added to the shader registry
for a given shader prim, assuming it is found in a shader definition file found by an
Ndr discovery plugin.

(Internal change: 1894015)
(Internal change: 1894055)
(Internal change: 1894094)
(Internal change: 1894372)