[Hd, HdSt] Added topology visibility operations for meshes in Hydra This change adds invisiblePoints and invisibleFaces to HdMeshTopology, which allows us to express topological visibility operations for meshes via the DirtyTopology change tracker bit.
Invisible points were handled in HdSt via a hack, wherein it came in as a sparse primvar, and was then expanded to a bitmask representation.

We don't want to recompute the topology buffers each time topological visibility changes since it's expensive.
To keep things fairly decoupled and simple, instead of adding topological visibility data to the topology BAR as "optional fields", we introduce a topologyVisibility BAR that holds this information.
Changes:
- Added topologyVisibility BAR to HdDrawItem, and an additional index to HdDrawingCoord.
- Extended resource binder, code gen and batching code to use it.
- Removed old hacky code in HdStMesh for points visibility.
- Added _PopulateTopologyVisibility to HdStMesh, which creates/resizes the topologyVisibility BAR when it is authored/updated respectively.
- Added visibility.glslfx to park the decoding logic for the topology visibility buffers, which is then used by the FS main in mesh.glslfx.

(Internal change: 1894058)