Improve the Renderer submenu in the case of unsupported renderers.
If an attempt to switch to a particular renderer fails, an error message
is set in the status bar. Then the offending renderer is disabled in
the renderer menu and has "(unsupported)" added to the menu item.

Remove code that checks if each renderer is supported before adding it
to the renderer menu. This involved loading all renderer plugins before
starting usdview which is not an expense the user should pay at startup.