[hd/hdEmbree/usdImagingGL] Build out the hydra AOV API, and add switchable AOVs to usdview (for supported backends).
AOVs have been added as a concept in both HdxTaskController (with a fairly rich API) and UsdImagingGL (with a much more limited API).  The usdview UI has been extended to add a "Hydra AOVs" menu, which lets the user choose which AOV is displayed.

To introduce this, HdxColorizeTask was introduced: it takes a renderbuffer and semantic, and produces color output.

(Internal change: 1894403)
(Internal change: 1894453)