[Hd, HdSt] Add Support for inverting the depth test of a draw target
For some multi-pass techniques, it is desirable to write the furthest prims to the output buffer instead of the typical nearest prims.
This change adds a control to Draw targets, that allows the depth test used by the draw targets to be inverted.

The depth compare function is still set up by the Draw Target Task parameters. However, draw targets now have a Depth Priority parameter, that will invert the direction of the task specified function. For example, Less Than or Equals becomes Greater Than. Equals, Not Equals, Always and Never are left untouched.

Changes:
- Added new HdDepthPriority enum to hydra.
- Added depth priority to HdStDrawTargetAttachmentDescArray with getter and setter.
- Added depth priority to HdStDrawTargetRenderPassState with getter and setter.
- HdStDrawTarget set's the value in the render pass state based on the description array.
- HdxDrawTargetTask resolves the comparison function between the task param and the priority specified in the pass state. It uses this information to setup the value in the HdStRenderPassState.
- Removed redundant setting of the glDepthFunc(). This is done by HdStRenderPassState::Bind().

(Internal change: 1894544)