[px_vp20] re-implement legacy viewport selection matrices getter to resemble Viewport 2.0 version
This change replaces the px_LegacyViewportUtils::GetViewSelectionMatrices()
function with a GetSelectionMatrices() that computes selection matrices the
same way that the Viewport 2.0 version of the function does. This removes
the need to use OpenGL through the M3dView to determine the selection
matrices.

The legacy viewport version of the batch renderer's TestIntersection()
method was modified to take an MSelectInfo instead of an M3dView to enable
this, and *both* viewport versions of TestIntersection() were changed to
no longer take singleSelection as a parameter, since it can just be read
from the selection context passed in.

The proxy shape selection performance tests were also adjusted to ensure
that the window that the test viewport is sized correctly to get us the
viewport size that we want. Previously, we were requesting a 960x540 window
resulting in a smaller viewport since we were not accounting for the window
frame and toolbar in the view widget. Now we should be getting a viewport
of exactly 960x540.

(Internal change: 1894592)