Move usdVolImaging from plugin to lib
Although the adapters in usdVolImaging are plugins
to the usdImaging library, they are still considered
a core part of the usdImaging component in the same
way the usdVol schema library is considered a core
part of the usd component. So the lib directory is
the appropriate home for this library.

(Internal change: 1894957)