[Hdx] Remove HdxTaskController's seperate id render task.  Now, you pass in enableIdRender to the render params, and it re-uses the same render task.
Duplicating the tasks made the API confusing and introduced bugs.  The main argument for it is to optimize the case of switching between fullscreen ID and color renders quickly, but nobody actually does this; the only fullscreen ID render consumers are usdscenemetrics, which never renders color, and usdview, where it's a menu toggle.

(Internal change: 1899850)