[Hd/HdSt/Hdx] Prevent aggregation of buffers with different resize frequencies.
This bug was a report of a poor performing shot when it contained an asset with a varying topology.  While varying topology is known to be an expensive change for an rprim, the performance profile showed this was scaling by the total number of Rprim in the scene.  Therefore, it is desirable to isolate these Rprims so they have less impact on Rprim that don't change.

The cause of the scalability issues is two-fold:
- Deep batch validation, which asses all draw batches to see if they need re-rebuilding
- Batch building: The cost of rebuilding the batch that contains the Rprim with the varying topology

This change focuses on the batch rebuilding issue.  Naturally, this algorithm scales based on the number of prims in the batch.  In the test shot, the prim with the varying toplogy was placed in a batch with a very large number of other prims, even though those prims are not changing.  So the batch builder was iterating over this large set of prims.

The reason why the batches got rebuilt is because the buffer array contain the topology buffer array range (BAR) got reallocated.  So this optimization focuses on splitting the buffer arrays, such that prims that are static are not in the same buffer array as prims that are dynamic.  This also has the added advantage of smaller buffer array reallocs as well, as only the buffers containing dynamic prims will get resized.

To enable this and possible future optimization based on update frequency, a new BAR property is being introduced called a usage hint.  This usage hint is a set of flags for communicating to the memory manager intended usage of the BAR, such as in this case, the size of the BAR will change over time.  There was a similar concept introduced for primvar sharing defining an "immutable" property.  This property has been integrated with the usage hints.

To determine if two BARs are compatible for aggregating into the same buffer array, an aggregation id is created for the BAR, with each memory manager having its own policy on generating the id (typically either an incremental value for simple memory managers that don't support aggregation or a hash of the BAR state).  For those that use the property hash, the usage hint has been added to that hash.

This different aggregation ids separates the varying prim from the non varying prim into different buffer arrays.   As because the dispatch system needs to bind different buffer arrays for these prims, the batches also get split, resulting in smaller batch rebuilds giving the speedup.

Changes:
- Added HdBufferArrayUsageHint bitfield structure with immutable and sizeVarying bits.
- Replaced BufferArray immutable property with the usage hint.
- Added the usage hint to the HdAggregationStrategy::ComputeAggregationId() api.
- HdResourceRegistry's Allocate*BufferArrayRange methods, now take the usage hint and pass it to the buffer array registry.
- HdStVBOMemoryManager's no longer needs to store a immutable property, as that is set through the allocate methods.
- MergeBufferArrayRange (and associated Merge*BufferArrayRange) now take the usage flags and will perform BAR migration if the usage hint changes.
- Memory Managers have been updated for new API containing the usage hint.
- Lots of users of the resource registry updated to pass a default (0) usage hint.
- HdSt's Mesh updated to detect varying topology and varying vertex primvar sizes and set the size varying bit.

Notes:

- Other prim types than Mesh and other Primvar frequency are believed to benefit from the hint, but will be addressed separately.

(Internal change: 1900556)