[pxrUsdMayaGL] push check for view object filtering down into PxrMayaHdShapeAdapter::_GetVisibility()
It may be desirable to test for other view-related state when computing
visibility for a shape, so this change modifies the signature for
UpdateVisibility() to take the entire M3dView object rather than just an
MSelectionList of the view's filtered object list. Since we also have a
display filter for USD proxy shapes, the UpdateVisibility() method of its
shape adapter class is a good place to check the state of that filter to
ensure that Hydra doesn't draw those types of shapes if they are being
filtered out of the view.

(Internal change: 1901222)