[pxrUsdMayaGL] add shape adapter function for converting display style and status to an HdReprSelector
This is an incremental refactoring change in the direction of moving
display style handling out of the "prepare" phase and into the "draw" phase
of viewport rendering. The display style is a property of the view/frame
context rather than of the object(s) being rendered, so it can (and does)
vary per viewport. That means it must be considered at every draw call
rather than during the prepare phase, which may only happen once per object
for multiple viewports with different styles. Display status, however, *is*
a property of objects and needs to be used in combination with the display
style to determine the repr.

This change also fixes handling of the "shade only active" viewport modes.

(Internal change: 1901746)