[Glf] Fix for crash in Texture Registry when plugin loading fails.
A recent build configuration issue was causing texture plugins to fail to load.  While that error was handled, it caused null texture handles to be placed in the texture registry.  The method GlfTextureRegistry::GarbageCollectIfNeeded() is not expecting that, so was crashing.

While, the assumption is correct, the changes adds some defensive code to GlfTextureRegistry::GarbageCollectIfNeeded() to protect against this condition.

The change also fixes up the cases where the null handle was being added to the registry, so that assumption remains true.

The recently added override for GetTextureHandle() was also missing handling some nullptr return cases from factory and texture handle, so
the error handling pattern has been applied to this case too.

Changes:
- In the two GlfTextureRegistry::GetTextureHandle()'s check _CreateTexture() didn't return null before adding to the registry
  (the 3rd overload doesn't use _CreateTexture() and already has a null handle check).
- In GarbageCollectIfNeeded() add a TF_VERIFY() that the handle isn't null.
- The new GetTextureHandle() override, now also calls it's own new versions of  _CreateTexture()  and adds the null return handling

(Internal change: 1902100)