[HdSt] Further fixes for performance regressions caused by size-varying optimization.
After the previous fix, two more cases have been discovered by performance testing.

1) Refinement can cause the buffer of elements in a BAR to be expanded from the number of points in the topology.
2) Primvars transitioning from Immutable to Size Varying were treated as new shared primvars.

The previous check to see if the vertex primvars were size varying used to compare the number of elements in the Bar to the number of points in the topology.  This check is only be valid is the primvars don't undergo some computation that expands the number of points first, for example refinement.

Therefore, if the scene contained a refined prim, when that prim was first munged, it would transition to size varying and the performance tests were picking up this transition happening.

It was felt that deeper discussions about how to handle this expansion were needed, so in order to capture the spirit of the original optimization, the mesh now has a flag for if we detect a varying topology and uses this to set the size varying hint on the vertex primvar instead.

The second issue was detected because of the first issue.  In the tests the primvars were shared primvars.  In the test, when they migrated to size varying, the hints were different, so the code thought they were "new" shared primvars, instead of transition to an mutable buffer.

Changes:
- Added _hasVaryingTopology flag to HdStMesh.
- _hasVaryingTopology is set when a changed topology is detected in HdStMesh::_PopulateTopology()
- HdStMesh::_PopulateVertexPrimvars() now uses _hasVaryingTopology instead of element count detection for setting sizeVarying buffer hint.
- Setting sizeVarying also clears the immutable flag.
- If the immutable flag is cleared, the code now migrates the buffer to a mutable buffer rather than creating a new immutable buffer.

(Internal change: 1902124)