[HdSt] Only do deep batch validation if the geometric shader really changed.
This bug was a report of a poor performing shot when it contained an asset with a varying topology.  While varying topology is known to be an expensive change for an rprim, the performance profile showed this was scaling by the total number of Rprim in the scene.  Therefore, it is desirable to isolate these Rprims so they have less impact on Rprim that don't change.

The cause of the scalability issues is two-fold:
- Deep batch validation, which asses all draw batches to see if they need re-rebuilding
- Batch building: The cost of rebuilding the batch that contains the Rprim with the varying topology

This change focuses on 1 issue that was triggering the deep validation: The geometric shader.

The topology scheme, has an impact on the refine level of a prim and therefore can result in geometric shader changes.

Therefore, Stream needs to inform the scene delegate that if topology changed, is may be requesting geometric shader state.  This needs to be done before Stream knows what has actually changed in the topology, so the topology dirty bit gets propagated to the display style dirty bit, which causes the geometric shader to be re-evaluated.

Now in this case, the subdivision scheme in the topology has not changed, so the re-evaluated geometric shader turns out to be the same as the original.  However, batches were always marked dirty, triggering the deep validation.

This simple optimization, therefore just avoid triggering this deep validation if the re-evaluated geometric shader matches the current geometric shader.

Changes:
- Surrounded the code in_UpdateDrawItemGeometricShader that updates the draw item and marks batches dirty with a check to see if the computed geomShader matches what is already in the draw item.

(Internal change: 1902381)