[pxrUsdTranslators, usdMaya] Cleanup instancerTranslate and translate pairs that cancel each other out
Most (but not all) of the time, the instancerTranslate op that gets written by the instancer
prim writer gets cancelled out by an opposite translate op that's written by the prototype prim
writer. Detect those cases in the instancer prim writer's PostExport and clean up the redundant
translate ops.

Note that this has to be done _after_ writing all of the prototypes because there are some
weird cases (namely involving pivots) where the instancerTranslate and translate ops won't
cancel each other out; the textUsdExportPointInstancer has an example of such a case.

(Internal change: 1902417)