Remove additional logic in UsdImagingGL::GetResourceAllocation().
Note that this means that the "Ref" engine will now just output empty
dictionaries, which is a change in behavior.  We think that's okay
because:

a) the ref engine doesn't support textures anyways
b) the ref engine is not meant to be used for performance tuning.

Also, updated usdview to look for textureMemory instead of
textureMemoryUsed as the latter was already being computed in the
hdEngine implementation.

(Internal change: 1902465)