[HdSt] Prevent deep validation when bar aggregation hasn't changed.
This bug was a report of a poor performing shot when it contained an asset with a varying topology. While varying topology is known to be an expensive change for an rprim, the performance profile showed this was scaling by the total number of Rprim in the scene. Therefore, it is desirable to isolate these Rprims so they have less impact on Rprim that don't change.

The cause of the scalability issues is two-fold:
- Deep batch validation, which asses all draw batches to see if they need re-rebuilding
- Batch building: The cost of rebuilding the batch that contains the Rprim with the varying topology

This change focuses on one of the causes of deep batch validation.

All batches always go through a process of shallow validation, which is checks the versions of the buffer array associated with the first draw item in the batch for changes.  This can detect when buffer array ranges (BARs) have been added or removed from the buffer array that cause it to be reallocated.

It is possible that a draw item to be changed to a new BAR that is no longer part of the original Buffer Array.  However, as BARs can be shared between draw items, the original BAR may still have references, thus not causes the Buffer Array to be reallocated, so it doesn't change.

To detect this case, a process of Deep Validation is need, which will examine each draw item within a batch to check that it is still part of the same buffer array.  Hydra doesn't currently have a per-batch selection of the type of validation needed, so when deep-validation is set, it is done for all batches.  This means the check is down for every draw items in the scene, so can get to be very expensive, even for largely untouched scenes.

This optimization is looking at that case when something happens that causes a prim to change to a new buffer array range (BAR). This happens when the bufferspec changes or the usage hint changes. So when this happens, the prim allocates a new bar and removes the reference to the old bar.

In the test shot, deep validation was being triggered, but no actual changes were needed.  This is because the new BAR was aggregated with the original buffer array, so the Buffer Array version changed and shallow validation catches the modified batch.

To resolve this issue, the deep validation is now only trigger if the new BAR is aggregated to a different buffer array than the original BAR.

Changes:
- In HdStMesh::_PopulateTopology(), If the range has changed, only mark batches dirty, if the new BAR is not aggregated with the old BAR.
- A similar change has been made to HdStMesh::_PopulateVertexPrimvars().

(Internal change: 1902914)