[Hd, UsdImagingGL] Add a "render settings" API to the HdRenderDelegate class.
There have been some linked requests related to setting out-of-band parameters on hydra backends.

This change adds the APIs:
- void HdRenderDelegate::SetRenderSetting(TfToken, VtValue)
- VtValue HdRenderDelegate::GetRenderSetting(TfToken)

... which can be used to pass this sort of data.  Since there's no schema, frontends and backends both need to opt into a set of common keys to make this system work, and backends should be robust to entries not being set.

Additionally, as a way for backends to communicate with UI about what user-facing settings they support, this change adds the API:
- HdRenderDelegate::GetRenderSettingsList()

This provides a list of reflection metadata on the settings that can be used to auto-populate a UI, or otherwise programmatically include/exclude UI elements.

As an example, HdEmbree now exports a small selection of render settings, and the usdview "Hydra Settings" menu is now populated from the backend-specific settings list.

(Internal change: 1902946)
(Internal change: 1902983)