When packaging assets into a usdz file, handle the following cases properly with appropriate fixes:
   * self-references - these are now converted to relative references and directory
   remapping is skipped for these.
   * references to the root asset layer that will be renamed in the end and references
to the original asset from the flattened asset (in the case of --arkitAsset). The remapping
function is now aware of the paths to the original root asset layer and the flattened root
asset layer and performs the correct remapping.
   * package dependencies - entire packages are copied instead of attempting to
export just the layer within a package that is a dependency.

(Internal change: 1903857)