[usdMaya] add UsdMayaUtil::GetMayaNodeName() function for getting a Maya node's name
Some places in the usdMaya code currently work with MDagPaths just so that
they can easily use its fullPathName() method to include node names in
warning or error messages if something goes wrong. In many cases, they do
not actually care or require that the node is a DAG node and may otherwise
be able to support DG nodes as well, which cannot be represented by an
MDagPath.

This change adds a utility function for getting the name of a node given
just its MObject. If it is a DAG node, the full path name is returned using
the MFnDagNode function set. If it is a DG node, the absolute name is
returned using the MFnDependencyNode function set. Hopefully this will open
up possibilities for existing code to be expanded to support both DAG and
DG Maya nodes.

(Internal change: 1903991)