[usdMaya] Fix problem where reference assemblies didn't draw in "playback" representation.
"Fast playback" mode is intended to force-skip the bound computation, since the bounds
can change from frame-to-frame. Semantically, this means that the bounding box is
"infinite", and not "empty". But UsdMayaProxyShape::boundingBox() returned MBoundingBox(),
which is an empty bounding box. Some of the VP2 imaging code upstream was now depending
on that, so proxy shapes in "fast playback" mode (i.e. ones generated by "playback" representation)
were not drawn.

Fixes #675

(Internal change: 1904655)