Improved playhead scrubbing experience in usdview.
When a user clicks at a point on the slider that corresponds to a certain frame,
the play head now moves to the frame instead of jumping 10 steps in that direction.
This behavior is implemented in a custom derived class of QSlider (named FrameSlider).
Most of the frame slider functionality has been moved out of appController into this
separate module.

(Internal change: 1905609)