Removing hdEngine.cpp, finally folding it into UsdImagingGLEngine.
This is the latest step in eventually being able to get rid of the legacy
implementation.  As it stands now, and probably for a little while longer,
it remains as an extra detail off to the side, and no longer part of the
design of the usdImagingGL library.

(Internal change: 1905756)