Fixed symbol export issues in Hd
HdRenderIndex and HdPerfLog's default constructor are not exported (no HD_API tag).
This causes a problem when a user tries to make a windows DLL since the header is trying to use a non-exported function.

Added exports for HdPerfLog's c'tor and d'tor and moved
the implementation of HdRenderIndex::New to the cpp file.

(Internal change: 1911580)
(Internal change: 1911891)