Don't check for cycles when adding variant arcs
Variant arcs represent the selection of a particular
branch of scene description; they are not sites in
the composed scenegraph that could cause a cycle.
For example, adding a child variant arc /A{var=sel}
to parent /A should not be counted as a cycle, even
though the child path is prefixed by the parent.

The cycle detection code had logic to handle the
above case, but it was erroneously flagging cycles
when a prim spec inside a variant contained an
internal sub-root reference to a prim that had
a spec inside the same variant. In the process of
building up ancestral opinions for the index for
the referenced prim, the _HasAncestorCycle function
would see a child site like /A{var=sel} and a parent
site like /A{var=sel}Child. Because the parent path
was prefixed by the child path, it would flag this
as a cycle, even though that child site would
eventually become /A{var=sel}Child2 in the final
referenced prim index.

Fixes #677

(Internal change: 1911634)