[Hd] Performance optimization for animated textures + large number of invisible prims
Animation was seeing slow playback when using animated textures, however certain operations were enough to clear this slow playback problem.

The cause of the slow playback, was having a very large number of prims (from a set) in the dirty list marked with the material id dirty and the varying bit is set.

This state was set, because the shot created a material with a texture that was a ptex after the set was loaded and marked invisible.  Due to the ptex transition, all Rprims had there material id marked dirty and the varying bit set.  However, as the set prims were invisible, this state was never cleared.

To optimize this, 3 operations have been added to hydra's change tracking and sync processing:
- Optimize redundant dirty state sets, to avoid entering the time-varying state.
- Transition invisible prims out of time-varying state when synced.
- If there are a lot of non time-varying prims, rebuilding the dirty list set to only include the time-varying prims.

The point of the first optimization, is to avoid thrashing the varying state on the invisible prims, when if the prim was visible it would be time-varying.

The second operation, removes invisible prims from the time-varying state, so that the dirty list gather operation will ignore this prims when entering the stable state.

The third operation detects the condition initially described and causes the dirty list to be rebuilt without the invisible prims.

Note, that there is a bit of a frame delay between these operations.
- At time 0, a delegate or hydra through propergation marks an invisible prim dirty.  That prim will have its varying state set.
- At 1st sync: We detect the varying state has changed, so we rebuild the dirty list including the invisible prims.
- During pre-sync the invisible prim is detected and it varying state is cleared.
- During 2nd sync: We detect there are too many invisible prims in the dirty list, so we trigger a rebuild of the varying state set on next sync.
- During the 3rd sync: We rebuild the dirty list, so Hydra returns to stable operation.

Adding prims to the render index is another case that doesn't set the varying state bit, so that can hit the 3rd operation and cause the varying state to be rebuilt.  While this does clear the varying bit on prims, the dirty list itself would have been transition to the initial state by adding the prims and the next sync would be building a new varying list anyway.  So is normal operation, this shouldn't cause an additional rebuild.

In order to debug these issues, I've added some addition debug output, with the HD_VARYING_STATE flag and a new debug flag HD_DIRTY_ALL_COLLECTIONS.
- HD_VARYING_STATE will now report a prims transitions into and out of the varying state.
- Marking all collections dirty is a heavy performance hammer and one we often need to debug.  HD_DIRTY_ALL_COLLECTIONS makes this easier by reporting the stack trace of when the collections are marked dirty, so it is easier to track down the cause.

Changes:
- In HdChangeTracker::MarkRprimDirty(), it will now early exit as a no-op, if the requested dirty bits are already marked dirty on the prim.
- Similarly HdChangeTracker::MarkAllRprimsDirty(), has been optomized to not introduce the varying state or increment the varying state version if the dirty bits request are already marked dirty.
- Added new function, ResetRprimVaryingState, to HdChangeTraker, that will clear the varying state on a prim.
- Added a stack dump to HdChangeTracker::MarkAllCollectionsDirty(), if the HD_DIRTY_ALL_COLLECTIONS debug flag is set.
- Added IsVarying query to HdChangeTracker.
- Added new HD_DIRTY_ALL_COLLECTIONS debug flag
- In _PreSyncRPrims(), if HdRprim::CanSkipDirtyBitPropagationAndSync() returns true, clear the varying state from the prim.
- In the loop that builds the sync map, fixed a case where int was still being used for HdDirtyBits.
- As well as looking for the number of skipped prims, because they are clean, it also looks at the number of non-varying prims in the list.

(Internal change: 1912157)